finish delete worker

+++ database/migrations/2017_04_12_094248_create_workers_table.php
@@ -18,7 +18,7 @@ class CreateWorkersTable extends Migration
             $table->integer('pid')->unsigned()->default(0);
             $table->tinyInteger('post_id')->unsigned()->default(0);
             $table->string('name', 100);
-            $table->smallInteger('salary')->unsigned()->default(0);
+            $table->integer('salary')->unsigned()->default(0);
             $table->string('avatar')->default('default.jpg');
             $table->timestamps();
         });
